+++ PlatformIntrinsics.h
@@ -48,11 +48,11 @@ void Unaligned_Mat4x4Mul(const float* a, const float* b, float* result);
 
 void Unaligned_RGBAToBGRA(uint32* image, size_t width, size_t height);
 
-void Aligned_Mat4x4Transform(const float matrix[4][4], float* data, size_t stride, size_t length);
+void Aligned_Mat4x4Transform(const float matrix[4][4], float* data, int32 stride, int32 length);
 
 void Aligned_DepthBufferVisualize(float* buffer, size_t width, size_t height);
 
-void Aligned_Vec4Homogenize(float* data, size_t stride, size_t length);
+void Aligned_Vec4Homogenize(float* data, int32 stride, int32 length);
 
 void Unaligned_BlendBuffers(uint32* devBuffer, uint32* frameBuffer, size_t width, size_t height, float opacity);
 
@@ -60,10 +60,10 @@ void Aligned_BackfaceCull(IndexBuffer& indexBuffer, const VertexBuffer& vertexBu
 
 void Unaligned_AABB(const float* vertices, size_t numVertices, size_t stride, float outMin[4], float outMax[4]);
 
-void Unaligned_FlatShadeRGB(const float* vertices, const size_t* indices, const size_t end,
+void Unaligned_FlatShadeRGB(const float* vertices, const int32* indices, const int32 end,
   const float maxWidth, const float maxHeight, uint8* framebuffer, float* depthBuffer);
 
-void Unaligned_FlatShadeUVs(const float* vertices, const size_t* indices, const size_t end,
+void Unaligned_FlatShadeUVs(const float* vertices, const int32* indices, const int32 end,
   const float maxWidth, const float maxHeight, uint8* framebuffer, float* depthBuffer, const Texture* texture);
 
 }
